Organize Code of Product Routes
