+++ Week3/Hackathon/backend/middlewares/productValidate.js
@@ -1,0 +1,13 @@
+import { validationResult } from "express-validator";
+
+export const validate = (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({
+      success: false,
+      message: errors.errors[0].msg,
+    });
+  }
+
+  next();
+};
